[Feat]: CouponEventWebController 병합
